+++ services/audio_suite/client/utils/src/audio_suite_eq_algo_interface_impl.cpp
@@ -184,10 +184,10 @@ int32_t AudioSuiteEqAlgoInterfaceImpl::SetParameter(const std::string &sEQLGain,
 
     stData_.piDataIn = reinterpret_cast<IMEDIA_INT32 *>(dataIn_.data());
     stData_.piDataOut = reinterpret_cast<IMEDIA_INT32 *>(dataOut_.data());
-    stData_.iSize = nodeParameter_.frameLen;
+    stData_.iSize = static_cast<int32_t>(nodeParameter_.frameLen);
     stData_.iEnable_SWS = 1;
     stData_.iData_Format16 = 1;
-    stData_.iData_Channel = nodeParameter_.inChannels;
+    stData_.iData_Channel = static_cast<int32_t>(nodeParameter_.inChannels);
     stData_.iMasterVolume = MASTERVOLUME;
     AUDIO_INFO_LOG("iMedia_SWS_DATA stData_ Init Success");
     return SUCCESS;
@@ -227,7 +227,7 @@ int32_t AudioSuiteEqAlgoInterfaceImpl::Apply(std::vector<uint8_t *> &pcmInBuf, s
     for (size_t i = 0; i < frameLen_; i++) {
         pcmOut[i] = ((unsigned int)dataOut_[i] >> TWO_BYTES_WIDTH);
     }
-    
+
     return result;
 }
 }  // namespace AudioSuite
